cambiata do_pivot aggiunta src/ , e creata get_rid_alphas in testing

+++ src/withparam.cpp
@@ -12,14 +12,14 @@ int withparam(int argv, char** argchar){
     int c,r;
     in>>r;
     in>>c;
-    float **A;
-    float *b;
-    float *f;
-    f = new float[c];
-    b = new float[r];
-    A = new float*[r];
+    double **A;
+    double *b;
+    double *f;
+    f = new double[c];
+    b = new double[r];
+    A = new double*[r];
     for(int i=0;i<r;i++){
-        A[i]=new float[c];
+        A[i]=new double[c];
         for(int j=0;j<c;j++){
             in>>A[i][j];
         }
